% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_oneshot_spatial.R
\name{analyse_oneshot_spatial}
\alias{analyse_oneshot_spatial}
\title{Use the Biodiverse server to load a set of rasters and then analyse them}
\usage{
analyse_oneshot_spatial(
  raster_files = NULL,
  r_data = NULL,
  spreadsheet_data = NULL,
  delimited_text_file_data = NULL,
  shapefile_data = NULL,
  cellsizes,
  calculations = c("calc_richness", "calc_endemism_central"),
  tree = NULL,
  ...
)
}
\arguments{
\item{raster_files}{character}

\item{r_data}{list}

\item{spreadsheet_data}{list}

\item{delimited_text_file_data}{list}

\item{shapefile_data}{list}

\item{cellsizes}{numeric}

\item{calculations}{character}

\item{tree}{class phylo}

\item{...}{passed on to start_server call}
}
\value{
The results of the analysis as an R list
}
\description{
Use the Biodiverse server to load a set of rasters and then analyse them
}
\examples{
if(interactive()) {
  analysis_results = analyse_oneshot_spatial (
    raster_files             = c("r1.tif", "r2.tif"),
    r_data                   = list(
                                    '250:250' = list (r1 = 13758, r2 = 13860),
                                    '250:750' = list (r1 = 11003, r2 = 11134),
                                   )
    spreadsheet_data         = c("r1.xlsx", "r2.xlsx"),
    delimited_text_file_data = c("r1.csv", "r2.csv"),
    shapefile_data           = c("r1.shp", "r2.shp"),
    cellsizes                = c(500, 500),
    calculations             = c("calc_endemism_central", "calc_richness", "calc_pd"),
    tree                     = some_phylo_tree
  )
}
}
