% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_rasters_spatial.R
\name{analyse_rasters_spatial}
\alias{analyse_rasters_spatial}
\title{Use the Biodiverse server to load a set of rasters and then analyse them}
\usage{
analyse_rasters_spatial(
  raster_files,
  cellsizes,
  calculations = c("calc_richness", "calc_endemism_central"),
  tree = NULL,
  ...
)
}
\arguments{
\item{raster_files}{character}

\item{cellsizes}{numeric}

\item{calculations}{character}

\item{tree}{class phylo}

\item{...}{passed on to start_server call}
}
\description{
Use the Biodiverse server to load a set of rasters and then analyse them
}
\examples{
if(interactive()) {
  analyse_rasters_spatial (
    raster_files = c("r1.tif", "r2.tif"),
    calculations = c("calc_endemism_central", "calc_richness", "calc_pd"),
    tree = some_phylo_tree
  )
}
}
