% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiodiverseR-package.R
\docType{package}
\name{BiodiverseR-package}
\alias{BiodiverseR}
\alias{BiodiverseR-package}
\title{BiodiverseR: An R interface to the Biodiverse engine for the spatial analysis of diversity}
\description{
BiodiverseR provides an interface to Biodiverse. Biodiverse is a tool for the spatial analysis of diversity using indices based on taxonomic, phylogenetic, trait and matrix-based (e.g. genetic distance) relationships, as well as related environmental variations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shawnlaffan/BiodiverseR}
  \item Report bugs at \url{https://github.com/shawnlaffan/BiodiverseR/issues}
}

}
\author{
\strong{Maintainer}: Shawn Laffan \email{shawn.laffan@unsw.edu.au} (\href{https://orcid.org/0000-0002-5996-0570}{ORCID})

Authors:
\itemize{
  \item Jakub Tomczyk \email{jstomczyk@gmail.com}
}

}
\keyword{internal}
