% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_service.R
\name{create_service}
\alias{create_service}
\title{Create service}
\usage{
create_service(path_net_str, path_service_str)
}
\arguments{
\item{path_net_str}{string, path_service_str string}
}
\description{
Create a service which allows us to access functionalities of the Biodiverse
tool.
}
\examples{
create_service("C:\\\\Windows\\\\Microsoft.NET\\\\Framework64\\\\v4.0.30319", "path")
}
