% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_perl_on_mac.R
\name{install_perlbrewr}
\alias{install_perlbrewr}
\title{Set Perl up on macOS}
\usage{
install_perlbrewr(perl_version = "5.36.1")
}
\arguments{
\item{perl_version}{character string specifying version of perl needed}
}
\description{
Set Perl up on macOS
}
