% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_perl_on_windows.R
\name{install_strawberry_perl}
\alias{install_strawberry_perl}
\title{Install a version of strawberry perl to run the server}
\usage{
install_strawberry_perl()
}
\description{
Install a version of strawberry perl to run the server
}
\examples{
if(interactive()) {
  install_strawberry_perl()
}
}
