% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data_}
\alias{load_data_}
\title{Load data onto the server associated with
a BiodiverseR::basedata object}
\usage{
load_data_(bd, params = list())
}
\arguments{
\item{bd}{class R6 basedata}

\item{params}{list}
}
\description{
Load data onto the server associated with
a BiodiverseR::basedata object
}
\examples{
if(interactive()) {
  b = BiodiverseR::basedata$new(name = "trial")
  success = b$load_data()
}
}
