% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_spatial_analysis.R
\name{process_tabular_results}
\alias{process_tabular_results}
\title{Process the results of the spatial analysis}
\usage{
process_tabular_results(call_results)
}
\arguments{
\item{call_results}{list}
}
\description{
Process the results of the spatial analysis
}
