% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cluster_analysis.R
\name{run_cluster_analysis}
\alias{run_cluster_analysis}
\title{Runs a cluster analysis with the given parameters
Uses the given basedata object to call the server}
\usage{
run_cluster_analysis(
  bd,
  index = "SORENSON",
  linkage_function = "link_average",
  calculations = NULL,
  spatial_conditions = NULL,
  def_query = NULL,
  name = NULL,
  tree = NULL,
  cluster_tie_breakers = NULL
)
}
\arguments{
\item{bd}{class R6 basedata}

\item{index}{character}

\item{linkage_function}{character}

\item{calculations}{character}

\item{spatial_conditions}{character}

\item{def_query}{character}

\item{name}{character}

\item{tree}{class phylo}

\item{cluster_tie_breakers}{character to handle tie breakers}
}
\value{
The processed results of the cluster analysis
}
\description{
Runs a cluster analysis with the given parameters
Uses the given basedata object to call the server
}
\examples{
if(interactive()) {
  bd = BiodiverseR::basedata$new(cellsizes=c(500,500))

  r_data = list(
                '250:250' = list (r1 = 13, r2 = 13),
                '250:750' = list (r2 = 11),
               )

  params = list (bd_params = r_data, raster_params = NULL)
  result = bd$load_data(params)

  results = bd$run_cluster_analysis (
    # calculations = calculations,
    # tree = tree
  )
}
}
