% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_randomisation_analysis.R
\name{run_randomisation_analysis}
\alias{run_randomisation_analysis}
\title{Runs a randomisation analysis by calling the server
using a basedata object}
\usage{
run_randomisation_analysis(
  bd,
  rand_function = "rand_structured",
  iterations = 999,
  name = NULL,
  spatial_conditions = NULL,
  def_query = NULL,
  ...
)
}
\arguments{
\item{bd}{class R6 basedata}

\item{rand_function}{character}

\item{iterations}{integer}

\item{name}{character}

\item{spatial_conditions}{character}

\item{def_query}{character}

\item{...}{}
}
\value{
Returns the results from calling the server
}
\description{
Runs a randomisation analysis by calling the server
using a basedata object
}
\examples{
if(interactive()) {
  bd = basedata$new(cellsizes=c(500,500))
  r_data = list(
                '250:250' = list (r1 = 13758, r2 = 13860),
                '250:750' = list (r1 = 11003, r2 = 11134),
               )

  params = list (bd_params = r_data, raster_params = NULL)
  bd$load_data(params)

  params = list (bd_params = r_data, raster_params = NULL)
  bd$load_data(params)

  bd$run_cluster_analysis (
    calculations = c("calc_endemism_central", "calc_richness", "calc_pd"),
    # tree = some_phylo_tree
  )

  bd$run_randomisation_analysis (
    bd                 = bd,
    rand_function      = 'some_rand_function',
    iterations         = integer,
    name               = 'some_name',
    spatial_conditions = c("calc_endemism_central", "calc_richness", "calc_pd"),
    def_query          = 'some_def_query'
  )
}
}
