% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_service.R
\name{create_service}
\alias{create_service}
\title{Create service}
\usage{
create_service(test_str)
}
\arguments{
\item{test_str}{string}
}
\value{

}
\description{
Create a service which allows us to access functionalities of the Biodiverse
tool.
}
\examples{
create_service("test")
}
